# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for the NutriMind nutrition and health tracking mobile app.",
        "title": "NutriMind API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/notifications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns paginated notification history for the authenticated user (sent push notifications), newest first.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "List notification history",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page size (default 10)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated notification list",
                        "schema": {
                            "$ref": "#/definitions/dto.ListNotificationsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/notifications/fcm-token": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Registers or updates the Firebase Cloud Messaging token for the authenticated user's device. Call this after sign-in and whenever the FCM token is refreshed by the OS.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Register FCM device token",
                "parameters": [
                    {
                        "description": "FCM token and platform",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.RegisterFCMTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token registered",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/reminders": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all reminder configs for the authenticated user (meal reminders, water reminders, etc.) including enabled state, frequency, and time windows.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reminders"
                ],
                "summary": "Get reminder configurations",
                "responses": {
                    "200": {
                        "description": "All reminder configs",
                        "schema": {
                            "$ref": "#/definitions/dto.GetRemindersResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/reminders/{type}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates or updates the reminder config for a given type (e.g. \"meal\", \"water\"). Supported fields: enabled, frequency_min, specific_times, window_start, window_end, custom_message.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "reminders"
                ],
                "summary": "Create or update reminder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Reminder type (e.g. meal, water)",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Reminder config",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpsertReminderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated reminder config",
                        "schema": {
                            "$ref": "#/definitions/dto.ReminderConfigResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.GetRemindersResponse": {
            "type": "object",
            "properties": {
                "reminders": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ReminderConfigResponse"
                    }
                }
            }
        },
...
        "dto.ListNotificationsResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.NotificationLogResponse"
                    }
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "dto.NotificationLogResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "deep_link": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "notification_type": {
                    "type": "string"
                },
                "scheduled_at": {
                    "type": "string"
                },
                "sent_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "dto.RegisterFCMTokenRequest": {
            "type": "object",
            "required": [
                "fcm_token",
                "platform"
            ],
            "properties": {
                "fcm_token": {
                    "type": "string"
                },
                "platform": {
                    "type": "string",
                    "enum": [
                        "android",
                        "ios"
                    ]
                }
            }
        },
        "dto.ReminderConfigResponse": {
            "type": "object",
            "properties": {
                "custom_message": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "frequency_min": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "reminder_type": {
                    "type": "string"
                },
                "specific_times": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "window_end": {
                    "type": "string"
                },
                "window_start": {
                    "type": "string"
                }
            }
        },
...
        "dto.UpsertReminderRequest": {
            "type": "object",
            "required": [
                "window_end",
                "window_start"
            ],
            "properties": {
                "custom_message": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "frequency_min": {
                    "type": "integer"
                },
                "specific_times": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "window_end": {
                    "type": "string"
                },
                "window_start": {
                    "type": "string"
                }
            }
        },
...
        "response.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.FieldError"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.Error"
                },
                "is_success": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT Bearer token — format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}